kitties loading to recycler by push
