--- app/src/main/java/com/example/sobes/CatData.kt
@@ -1,5 +1,0 @@
-package com.example.sobes
-
-data class CatData(
-    val url: String
-    )
